requirements.txt
# ── 核心运行时 ────────────────────────────────────────────────
openai>=1.0.0
schedule>=1.2.0
requests>=2.31.0

# ── 配置与环境 ────────────────────────────────────────────────
python-dotenv>=1.0.0
PyYAML>=6.0.0

# ── MCP Server ────────────────────────────────────────────────
mcp>=1.0.0

# ── 钉钉官方 SDK（主动 API / Stream 事件订阅，按需保留） ────────
# dingtalk-stream>=0.18.0
